requirements.txt
playwright==1.42.0
pandas==2.2.1
openpyxl==3.1.2
google-generativeai==0.4.0
pdfplumber==0.11.0
fpdf2==2.7.9
python-dotenv==1.0.1
